Move Spring Secutiry common things to common package
